refactor: entity package
