+++ app/src/main/java/com/example/wellfed/mealplan/MealBookFragment.java
@@ -35,8 +35,13 @@ import android.widget.TextView;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 
+import com.example.wellfed.MealPlanAdapter;
 import com.example.wellfed.R;
+
+import java.util.ArrayList;
 
 public class MealBookFragment extends Fragment {
     private TextView userFirstNameTextView;
